# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PayRaider API",
    "description": "API for Stellar network analytics, anchor monitoring, and payment corridor insights",
    "contact": {
      "name": "PayRaider Team",
      "email": "support@payraider.io"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://api.payraider.io",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/admin/ip-whitelist": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "GET /admin/ip-whitelist - List all whitelisted IPs",
        "operationId": "list_whitelist",
        "responses": {
          "200": {
            "description": "Whitelist entries returned"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied (IP not whitelisted)"
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "POST /admin/ip-whitelist - Add IP to whitelist",
        "operationId": "add_to_whitelist",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddWhitelistRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Entry added to whitelist"
          },
          "400": {
            "description": "Invalid IP or CIDR"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied (IP not whitelisted)"
          }
        }
      }
    },
    "/admin/ip-whitelist/check": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "POST /admin/ip-whitelist/check - Check if an IP is whitelisted",
        "operationId": "check_whitelist",
        "requestBody": {
          "description": "IP address to check, e.g. {\"ip\": \"192.168.1.1\"}",
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Check result returned"
          },
          "400": {
            "description": "Invalid IP format"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/admin/ip-whitelist/{ip_or_cidr}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "DELETE /admin/ip-whitelist/{ip_or_cidr} - Remove IP from whitelist",
        "operationId": "remove_from_whitelist",
        "parameters": [
          {
            "name": "ip_or_cidr",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Entry removed from whitelist"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied (IP not whitelisted)"
          },
          "404": {
            "description": "Entry not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddWhitelistRequest": {
        "type": "object",
        "required": [
          "ip_or_cidr"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "ip_or_cidr": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}